Ajout Tabou

+++ Ordonnancement.py
@@ -56,7 +56,7 @@ class Ordonnancement:
         return self.getVoisCouple(jobsExchanges[0], jobsExchanges[1])
     
     def getAllVoisCouple(self):
-        lstAllVois = [None for _ in range(self.getNbJobs() * (self.getNbJobs() - 1) / 2)]
+        lstAllVois = [None for _ in range(int(self.getNbJobs() * (self.getNbJobs() - 1) / 2))]
         i = 0
         for i1 in range(self.getNbJobs() - 1):
             for i2 in range(i1 + 1, self.getNbJobs()):
@@ -76,7 +76,7 @@ class Ordonnancement:
         return self.getVoisInverseSeq(bornesSeq[0], bornesSeq[1])
     
     def getAllVoisInverseSeq(self):
-        lstAllVois = [None for _ in range(self.getNbJobs() * (self.getNbJobs() - 1) / 2)]
+        lstAllVois = [None for _ in range(int(self.getNbJobs() * (self.getNbJobs() - 1) / 2))]
         i = 0
         for i1 in range(self.getNbJobs() - 2):
             for i2 in range(i1 + 2, self.getNbJobs()):
@@ -97,7 +97,7 @@ class Ordonnancement:
         return self.getVoisPermSeq(bornesSeq[0], bornesSeq[1])
     
     def getAllVoisPermSeq(self):
-        lstAllVois = [None for _ in range(self.getNbJobs() * (self.getNbJobs() - 1) / 2)]
+        lstAllVois = [None for _ in range(int(self.getNbJobs() * (self.getNbJobs() - 1) / 2))]
         i = 0
         for i1 in range(self.getNbJobs() - 2):
             for i2 in range(i1 + 2, self.getNbJobs()):
